moq-gst: promote the rewritten core to the moqsink element
The parallel moqsinkspike element, built across the preceding commits, replaces
the original moqsink. Remove the old implementation, move the new core into
src/sink/, and register it as "moqsink" with the GType name kept as "MoqSink"
for registration continuity.

Behavior is a compatible superset of the old element: the same request sink
pads, the same seven codec caps, and the same three writable properties (url,
broadcast, tls-disable-verify), plus three read-only diagnostics (connected,
moq-version, estimated-send-bitrate).
